Quitado el jar de bets, y implementada casi toda la logica de BLFacade y los DAOs

+++ src/main/java/dev/sertis/betsjsf/dao/BetDAO.java
@@ -10,6 +10,6 @@ public interface BetDAO {
 
     void delete(Bet bet);
 
-    Bet getBetById(Integer id);
+    Bet getBetById(Long id);
 
 }
